Automatic File Naming
Set the output file to match our nomenclature. The file name will start with the folder name, then see if a take number is present in the audio source. If so, it'll increment it by 1. If not, then it'll set it to tk01. There is also a field now for Task Code which will add the production stage to the end of the file.